* net/loveruby/cflat/utils/TextUtils.java (toUnsigned): simplify.
git-svn-id: file:///Users/aamine/c/gitwork/public/cbc/trunk@3998 1b9489fe-b721-0410-924e-b54b9192deb8

+++ net/loveruby/cflat/utils/TextUtils.java
@@ -39,7 +39,7 @@ public class TextUtils {
     }
 
     static protected int toUnsigned(byte b) {
-        return (b >= 0) ? (int)b : 256 + (int)b;
+        return b >= 0 ? b : 256 + b;
     }
 
     static public boolean isPrintable(int c) {
